requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
autogen
python-docx
pypdf
easyocr
pdf2image
streamlit
reportlab
pymupdf
pytest